==> Jasen/1_Jasens_Workspace.Rmd <==
---
title: "Jasens_Workspace"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Helper function

```{r}
get_plots <- function(name,count, graphs){
  graph_count <- count
  index <- 0
  lay1 <- rbind(c(1, NA),
                c(NA, NA))
  lay2 <- rbind(c(1, 2),
                c(NA, NA))
  lay3 <- rbind(c(1, 2),
                c(3, NA))  
  lay4 <- rbind(c(1,2),
                c(3,4))
  
  pdf(name) 
  while(graph_count > 0){
    if(graph_count == 1){
      grid.arrange(graphs[[1+index]], layout_matrix = lay1)      
    } else if(graph_count == 2){
      grid.arrange(graphs[[1+index]], 
                   graphs[[2+index]], layout_matrix = lay2)      
    } else if(graph_count == 3){
      grid.arrange(graphs[[1+index]], 
                   graphs[[2+index]], 
                   graphs[[3+index]], layout_matrix = lay3)      
    } else{
      grid.arrange(graphs[[1+index]], 
                   graphs[[2+index]], 
                   graphs[[3+index]], 
                   graphs[[4+index]], layout_matrix = lay4)      
    }
    graph_count <- graph_count - 4
    index <- index + 4
  }
  dev.off()
}
```

# 0 count how many variables are missing

```{r}
df <- read.csv('cancer_registry.csv')

colSums(is.na(df))
```

# 0.5 Exploring insurance

```{r}
priv_vec <- df$PctPrivateCoverage - df$PctPrivateCoverageAlone
pub_vec <- df$PctPublicCoverage - df$PctPublicCoverageAlone
diff <- priv_vec - pub_vec
hm <- data.frame(priv_vec, pub_vec)

summary(diff)

all_races <- df$PctAsian + df$PctBlack + df$PctWhite + df$PctOtherRace

summary(all_races)

```

We realize we can easily impute for # college 18-24 since it needs to add to 100% with the other 18-24 groups.

## 1. Preliminary Regression

Here, I examine variables one at a time

```{r cars}
library(tidyverse)
library(gridExtra)
library(gplots)
df <- read.csv('cancer_registry.csv') %>%
  mutate(PctSomeCol18_24 = 100 - PctNoHS18_24 - PctHS18_24 - PctBachDeg18_24) %>%
  filter(incidenceRate < 1000) %>%
  filter(avgAnnCount < 20000) %>%
  filter(MedianAge < 200) %>%
  filter(AvgHouseholdSize > 1)

df <- na.omit(df)

p_values_1 <- c()
graphs1 <- list()
graph_count <- 1

vars <- colnames(df)
misc_vars <- c('binnedInc', 'Geography', 'TARGET_deathRate')
vars_1 <- setdiff(vars, misc_vars)
var_explained_1 <- c()

for(i in vars_1){
  eq <- paste('TARGET_deathRate', i, sep = ' ~ ')
  m1 <- lm(formula = eq, data = df)
  s1 <- summary(m1)
  
  # get p-value
  temp <- s1$coefficients[,4]
  temp1 <- unname(temp[2])
  p_values_1 <- append(p_values_1, temp1)
  
  var_explained_1 <- append(var_explained_1, s1$r.squared)
  coeffs <- as.numeric(m1$coefficients)
  x_temp <- dplyr::pull(df, i)
  y_hat <- x_temp  * coeffs[2] + coeffs[1]
  
  df_hat <- data.frame(y_hat, x_temp)
  g <- ggplot() + geom_point(data = df, aes(x = !!as.name(i), y = TARGET_deathRate)) + 
  geom_line(data = df_hat, aes(x = x_temp, y = y_hat))
  
  g
  graphs1[[graph_count]] <- g
  graph_count <- graph_count + 1
}

var_explained_df_1 <- data.frame(vars_1, var_explained_1, p_values_1)

get_plots('Jasen_1_Individual_Plots.pdf', graph_count - 1, graphs1)

```


# 2 Preliminary Log Regression

A few variables may look better with a log x axis

```{r}
log_vars <- c('avgAnnCount', 'avgDeathsPerYear', 'popEst2015', 'studyPerCap', 'PctBachDeg18_24', 'PctAsian', 'PctOtherRace')
log_names <- c()
graph_count <- 1
graphs2 <- list()
var_explained_2 <- c()
p_values_2 <- c()

for(i in log_vars){
  temp <- paste(i, '_log', sep = '')
  
  log_names <- append(log_names, temp)
  if(i %in% c('studyPerCap')){
    df <- df %>% mutate(!!as.name(temp) := log(!!as.name(i) + 1))
  } else if(i %in% c('PctAsian', 'PctBachDeg18_24', 'PctOtherRace')){
    df <- df %>% mutate(!!as.name(temp) := log(!!as.name(i) + exp(-5)))
  } else{
    df <- df %>% mutate(!!as.name(temp) := log(!!as.name(i)))
  }
  
  
  eq <- paste('TARGET_deathRate', temp, sep = ' ~ ')
  m2 <- lm(formula = eq, data = df)
  s2 <- summary(m2)
  
  # get p-value
  temp_p_value <- s2$coefficients[,4]
  temp_p_value_1 <- unname(temp_p_value[2])
  p_values_2 <- append(p_values_2, temp_p_value_1)
  
  var_explained_2 <- append(var_explained_2, s2$r.squared)
  coeffs <- as.numeric(m2$coefficients)
  x_temp <- dplyr::pull(df, temp)
  y_hat <- x_temp  * coeffs[2] + coeffs[1]
  
  df_hat <- data.frame(y_hat, x_temp)
  g <- ggplot() + geom_point(data = df, aes(x = !!as.name(temp), y = TARGET_deathRate)) + 
  geom_line(data = df_hat, aes(x = x_temp, y = y_hat))
  
  graphs2[[graph_count]] <- g
  graph_count <- graph_count + 1
}

get_plots('Jasen_2_Log_Individual_Plots.pdf', graph_count - 1, graphs2)


var_explained_df_2 <- data.frame(log_names, var_explained_2, p_values_2)
```
# 2.5 Displaying non-log and log of average death data

```{r}
g1_temp <- graphs1[[2]]
g2_temp <- graphs2[[2]]
lay <- rbind(c(1,2))

grid.arrange(g1_temp, g2_temp, layout_matrix = lay)
```

# 3 Look at all variables together

```{r}
dep_var <- c('TARGET_deathRate')
all_vars <- setdiff(vars_1, log_vars)
all_vars <- append(all_vars, log_names)
all_vars <- sort(all_vars)
eq3 <- paste(all_vars, collapse = ' + ')
eq3 <- paste(dep_var, eq3, sep = ' ~ ')

m3 <- lm(eq3, data = df)
s3 <- summary(m3)
s3

p_values <- unname(s3$coefficients[,4])
combined_p_values <- p_values[2:length(p_values)]
var_summary <- data.frame(all_vars, combined_p_values)

colnames(var_explained_df_2) <- colnames(var_explained_df_1)
var_explained_df_3 <- rbind(var_explained_df_1, var_explained_df_2)
colnames(var_explained_df_3) <- c('all_vars', 'individual_var_explained', 'individual_p_values')

var_summary <- var_summary %>% left_join(var_explained_df_3, by = 'all_vars')

var_summary2 <- var_summary %>% arrange(desc(individual_var_explained))

var_summary2
```


# 4 Heatmap

We have a decent ranking of the variables with criteria being p-value and % variance explained. We now look into variables that have high mutual correlation

```{r}

col1 <- c()
col2 <- c()
corr_col <- c()

for(i in all_vars){
  for(j in all_vars){
    col1 <- append(col1, i)
    col2 <- append(col2, j)
    x <- dplyr::pull(df, i)
    y <- dplyr::pull(df, j)
    
    temp_df <- data.frame(x,y) 
    temp_df <- na.omit(temp_df)
    
    temp_cor <- cor(temp_df$x, temp_df$y, method = 'pearson')
    corr_col <- append(corr_col, temp_cor)
  }
}

df4 <- data.frame(col1, col2, corr_col)

g4 <- ggplot() + geom_tile(aes(x = col1, y = col2, fill = corr_col)) + 
  theme(axis.text.x = element_text(angle = 90))


pdf('Jasen_heatmap1.pdf')
g4
dev.off()

#another heatmap option

df_temp <- df %>% select(all_vars)
df_temp <- na.omit(df_temp)

heatmap2 <- cor(df_temp)

pdf('Jasen_heatmap2.pdf')
heatmap.2(heatmap2, trace = 'none', margins = c(10,10), col = 'cm.colors', cexRow=0.7, cexCol = 0.7)
dev.off()

heatmap.2(heatmap2, trace = 'none', margins = c(10,10), col = 'cm.colors', cexRow=0.7, cexCol = 0.7)

```

# 5 Choosing variables solely on p-value

```{r}
choose_pvalue <- var_summary %>% filter(individual_p_values < 10^(-4)) %>%
  select(all_vars)

choose_pvalue <- unname(unlist(choose_pvalue))


eq5 <- paste(choose_pvalue, collapse = ' + ')
eq5 <- paste(dep_var, eq5, sep= ' ~ ')

m5 <- lm(eq5, data = df)
s5 <- summary(m5)
```

# 6 PCA of all 31 covariates

```{r}
num_comp <- 31

all_vars_and_dep <- append(all_vars, dep_var)
df6 <- df %>% select(all_vars_and_dep) 
df6 <- na.omit(df6)

df6_temp <- df6 %>% select(- !!as.name(dep_var))

m6 <- prcomp(df6_temp)
s6 <- summary(m6)

s6

component_matrix <- data.frame(as.matrix(df6_temp) %*% m6$rotation)
pca_df6 <- data.frame(component_matrix[,1:num_comp])

names <- c()
for(i in 1:num_comp){
  temp <- paste('PC', i, sep = '')
  names <- append(names, temp)
}

colnames(pca_df6) <- names

pca_df6 <- pca_df6 %>% mutate(TARGET_deathRate := df6$TARGET_deathRate)

eq6 <- paste(dep_var, ' ~ ', sep = '')
for(i in 1:num_comp){
  temp <- paste('PC', i, sep = '')
  eq6 <- paste(eq6, temp, sep = ' + ')
}
m61 <- lm(eq6, data = pca_df6)
s61 <- summary(m61)

# eq61 <- paste(names, collapse = ' + ')
# eq61 <- paste(' ~ ', eq61, sep = '')
# 
# pairs(pca_df6, eq61)
```

# 7 Manually choosing variables based on heatmap + p-value + percent variance explained

```{r}
vars_7 <- c('PctBachDeg25_Over',
            'PctPublicCoverageAlone',
            'incidenceRate',
            'medIncome',
            'PctEmployed16_Over',
            'PctMarriedHouseholds',
            'PctBlack',
            'PctOtherRace_log',
            'avgAnnCount_log')

eq7 <- paste(vars_7, collapse = ' + ')
eq7 <- paste(dep_var, eq7, sep = ' ~ ')

m7 <- lm(eq7, data = df)
s7 <- summary(m7)
```

# 8 PCA of these manually selected variables

```{r}
df8 <- df %>% select(vars_7) 
df8 <- na.omit(df8)
m8 <- prcomp(df8)
s8 <- summary(m8)

component_matrix <- data.frame(as.matrix(df8) %*% m8$rotation)
```
