==> Jasen/2_Jasen_Missing_Data.Rmd <==
---
title: "Jasen_Missing_Data"
author: "Jasen Zhang"
date: "11/23/2020"
output: pdf_document
---

# 1 Load The Data

```{r}
library(tidyverse)
library(gridExtra)
library(gplots)
library(glmnet)
library(mice)

dir = getwd()
data_dir <- paste(substr(dir,1, nchar(dir)-5), "cancer_registry.csv", sep = '')

df <- read.csv(data_dir) %>%
  mutate(PctSomeCol18_24 = 100 - PctNoHS18_24 - PctHS18_24 - PctBachDeg18_24) %>%
  filter(incidenceRate < 1000) %>%
  filter(avgAnnCount < 20000) %>%
  filter(MedianAge < 200) %>%
  filter(AvgHouseholdSize > 1)
```

# 2 Modifying the Design Matrix
- Getting rid of binnedInc and Geography

```{r}
vars <- colnames(df)
misc_vars <- c('binnedInc', 'Geography', 'TARGET_deathRate')
vars_1 <- setdiff(vars, misc_vars)

response_vars <- c('TARGET_deathRate')
predict_vars <- paste(vars_1, collapse = ' + ')

df <- df %>% select(- c('binnedInc', 'Geography'))
df <- df %>% mutate('ID' = rownames(df))

df <- data.frame(sapply(df, as.numeric))
```

# 3. Checking what columns have NA values

```{r}
colSums(is.na(df))
```

# 4. Imputing the two variables with NA

```{r}
imputed_df <- mice(df)
imputed_df2 <- imputed_df$imp

imputed_PctEmployed16_Over <- imputed_df2$PctEmployed16_Over
imputed_PctPrivateCoverageAlone <- imputed_df2$PctPrivateCoverageAlone

avg_imputed_PctEmployed16_Over <- data.frame(apply(imputed_PctEmployed16_Over, 1, mean))
avg_imputed_PctPrivateCoverageAlone <- data.frame(apply(imputed_PctPrivateCoverageAlone, 1, mean))

colnames(avg_imputed_PctEmployed16_Over) <- c('Imputed_PctEmployed16_Over')
colnames(avg_imputed_PctPrivateCoverageAlone) <- c('Imputed_PctPrivateCoverageAlone')

avg_imputed_PctEmployed16_Over <- avg_imputed_PctEmployed16_Over %>% mutate('ID' = as.numeric(rownames(avg_imputed_PctEmployed16_Over)))
avg_imputed_PctPrivateCoverageAlone <- avg_imputed_PctPrivateCoverageAlone %>% mutate('ID' = as.numeric(rownames(avg_imputed_PctPrivateCoverageAlone)))

df_imp <- df %>% left_join(avg_imputed_PctEmployed16_Over, by = 'ID') %>% 
  left_join(avg_imputed_PctPrivateCoverageAlone, by = 'ID')

df_imp[is.na(df_imp)] <- 0
```

# 5. Log transforming skewed variables

```{r}
df_imp2 <- df_imp
log_vars <- c('avgAnnCount', 'avgDeathsPerYear', 'popEst2015', 'studyPerCap', 'PctBachDeg18_24', 'PctAsian', 'PctOtherRace')
log_names <- c()


for(i in log_vars){
  temp <- paste(i, '_log', sep = '')
  
  log_names <- append(log_names, temp)
  if(i %in% c('studyPerCap')){
    df_imp2 <- df_imp2 %>% mutate(!!as.name(temp) := log(!!as.name(i) + 1))
  } else if(i %in% c('PctAsian', 'PctBachDeg18_24', 'PctOtherRace')){
    df_imp2 <- df_imp2 %>% mutate(!!as.name(temp) := log(!!as.name(i) + exp(-5)))
  } else{
    df_imp2 <- df_imp2 %>% mutate(!!as.name(temp) := log(!!as.name(i)))
  }
}

df_imp3 <- df_imp2 %>% select(- log_vars)
```

# 6. Add interaction terms, remove median age male and female

```{r}
df_imp4 <- df_imp3 %>% select(-c('MedianAgeMale', 'MedianAgeFemale')) %>%
  mutate(hs_black = PctHS25_Over * PctBlack) %>%
  mutate(hs_white = PctHS25_Over * PctWhite) %>%
  mutate(college_black = PctBachDeg25_Over * PctBlack)
```

# 7. Save the dataset

```{r}
data_dir <- paste(dir, '/df_imp.RData', sep = '')

save(df_imp4, file = data_dir)
```
