==> Jasen/6_Jasen_Final_Prediction_Results.Rmd <==
---
title: "Final Prediction Results"
author: "Jasen"
date: "12/5/2020"
output: pdf_document
---


```{r}
library(tidyverse)
library(gridExtra)
library(gplots)
library(glmnet)
```

# 1. Table of Coefficients

```{r}
final_coeffs <- merge(df_step_coeffs, lasso_final_df, by = "row.names", all = TRUE)
rownames(final_coeffs) <- final_coeffs$Row.names
final_coeffs <- final_coeffs %>% select(-c(Row.names))

#the loadings for the PCA
final_loadings

#the betas for the 10 PC's
summary(pca_train_model_2)
```

2. Table of R^2/RMSE

```{r}
R_squared_train <- c(r_squared_lasso_train, R_squared_PCA_train, R_squared_step_train)
R_squared_test <- c(r_squared_lasso_test, R_squared_PCA_test, R_squared_step_test)
RMSE_train <- c(rmse_lasso_train, RMSE_PCA_train, RMSE_step_train)
RMSE_test <- c(rmse_lasso_test, RMSE_PCA_test, RMSE_step_test)

final_metrics <- rbind(R_squared_train, R_squared_test, RMSE_train, RMSE_test)
colnames(final_metrics) <- c('Lasso', 'PCA', 'Stepwise')

train_metrics <- rbind(R_squared_train, RMSE_train)
colnames(train_metrics) <- c('Lasso', 'PCA', 'Stepwise')

test_metrics <- rbind(R_squared_test, RMSE_test)
colnames(test_metrics) <- c('Lasso', 'PCA', 'Stepwise')
```

3. All y_pred vs y_true graphs. 

```{r}
lay <- rbind(c(1,2,3),
             c(4,5,6),
             c(7,8,9))

lay1 <- rbind(c(1,2,3))


pdf('final_graphs.pdf', height = 15, width = 15)
grid.arrange(g_lasso_hyper,g_lasso_train,g_lasso_test,
             g_pca_hyper,g_pca_train,g_pca_test,
             g_step_hyper,g_step_train, g_step_test, layout_matrix = lay)
dev.off()

pdf('final_lasso_graphs.pdf', height = 5, width = 15)
grid.arrange(g_lasso_hyper,g_lasso_train,g_lasso_test, layout_matrix = lay1)
dev.off()

pdf('final_PCA_graphs.pdf', height = 5, width = 15)
grid.arrange(g_pca_hyper,g_pca_train,g_pca_test, layout_matrix = lay1)
dev.off()

pdf('final_stepwise_graphs.pdf', height = 5, width = 15)
grid.arrange(g_step_hyper,g_step_train,g_step_test, layout_matrix = lay1)
dev.off()

```
